scripts/requirements.txt
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
databricks-sdk==0.53.0
dotenv==0.9.9
google-auth==2.40.1
idna==3.10
numpy==2.2.6
pandas==2.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
rsa==4.9.1
six==1.17.0
tzdata==2025.2
urllib3==2.4.0
